Add support for array of values in 'Val' param
Value parameter will now accept an Array, which allows for
storage of multiple attributes from the initial objects. This was
accomplished using JSON.stringify to store objects in the
data-value attribute of the items.

Additional tests were added to cover the new usage, but I'm not
totally familiar with the bootstrap test suite so they may be
incomplete.